requirements.txt
requests
PyYAML
pandas
streamlit
plotly